📑 Allow page frontmatter in export article validation
