+++ .changeset/gold-bulldogs-carry.md
@@ -1,0 +1,5 @@
+---
+'myst-frontmatter': patch
+---
+
+Allow page frontmatter in export article validation
