+++ packages/myst-frontmatter/src/project/types.ts
@@ -4,7 +4,39 @@ import type { Licenses } from '../licenses/types.js';
 import type { Numbering } from '../numbering/types.js';
 import type { ProjectSettings } from '../settings/types.js';
 import type { SiteFrontmatter } from '../site/types.js';
+import { SITE_FRONTMATTER_KEYS } from '../site/validators.js';
 import type { ExpandedThebeFrontmatter } from '../thebe/types.js';
+
+export const PROJECT_AND_PAGE_FRONTMATTER_KEYS = [
+  'date',
+  'name',
+  'doi',
+  'arxiv',
+  'open_access',
+  'license',
+  'binder',
+  'source',
+  'subject',
+  'biblio',
+  'oxa',
+  'numbering',
+  'bibliography',
+  'math',
+  'abbreviations',
+  'exports',
+  'settings', // We maybe want to move this into site frontmatter in the future
+  // Do not add any project specific keys here!
+  ...SITE_FRONTMATTER_KEYS,
+];
+
+export const PROJECT_FRONTMATTER_KEYS = [
+  ...PROJECT_AND_PAGE_FRONTMATTER_KEYS,
+  // These keys only exist on the project
+  'references',
+  'requirements',
+  'resources',
+  'thebe',
+];
 
 export type ProjectAndPageFrontmatter = SiteFrontmatter & {
   date?: string;
